MM-65977: Add new MattermostPlugins setting (#948)
* Move ProvisionURL under Terraform struct

This is done mainly for having access to t.config.AWSAMIUser, so that we
can bulid the destination path of the provisioned file.

As a bonus, add such destination path as a return value, so that callers
know where it is located.

* Improve error logging when uploading a file

* Add a new MattermostPlugins setting

This new setting lets users specify the exact set of plugins to install
in the deployment, with the possibility to specify either officially
released plugins through their release URL, or locally built ones
through a local filesystem path.