Register plugins and their actions
Add a new plugins field to the SimulController struct, which contains
all the registered plugins, initialized in the New function using the
plugins.GeneratePluginControllers function.

We also make sure to call ClearUserData right after calling
User.ClearUserData, to reset the stores of both the regular controller
and the plugin's controller.