Implement the hooks glue code
Add new SimulController.RunHookXYZ functions, which range over the
initialized plugins, calling their respective HookXYZ functions inside
goroutines.

These new functions are called from the SimulController's actions points
where they need to be called: after logging in, switching channel or
switching team.

As a result, the switchChannelFunction needed to be added as a
SimulController method, instead of a standalone function, since it needs
to access the RunHookSwitchChannel function to call the plugins' hooks.